test: param_dtype propagation tests expect fp32 output (fp32-island endpoints)
Pre-existing pure-bf16-storage tests asserted bf16 model output; the shared
FinalPatchExpand norm is now a constructed-fp32 island, so models emit fp32.
This is the plan's end-state assertion, pulled forward to keep the suite green.

+++ tests/test_healconv.py
@@ -206,5 +206,6 @@ def test_healconv_param_dtype_propagates():
 
     x = jax.random.normal(jax.random.key(0), (2, p.npix, 3))
     y = model(x)
-    assert y.dtype == jnp.bfloat16 and y.shape == (2, p.npix, 5)
+    # models emit fp32: the output conv is an fp32 island regardless of knobs
+    assert y.dtype == jnp.float32 and y.shape == (2, p.npix, 5)
     assert bool(jnp.isfinite(y).all())
